AD-HOC feat (composer): Add a composer file specification
Composer is a package manager generally used by some of the more modern
PHP frameworks. It can be used in Magento 1 with the assistance of tools
that wire the composer packages in a specified format (described in a
modman file) to a Magento store. Such a package manager may be found
here:

https://github.com/Cotya/magento-composer-installer

It is the primary way of managing Magento packages in the authors
organisation. This commit adds such a composer file, with a series of
defaults that the author has previously found desirable