AD-HOC feat (head): Add a reference to the exposed manifest.json endpoint
Recent work has established an endpoint at manifest.json that can be
queried for certain offline information. However, at this time, the
browser has no knowledge of this endpoint.

This commit adds a layout file, and reference to this file such that the
`<head>` block will include a reference to the manifest.json.

This has been verified as a reasonable solution, based on the Chrome
LightHouse audit.