ISSUE-23 fix (theme): Correct the location of theme file
In previous work, a design file was added to indicate that manifest.json
could be found at the URL /manifest.json. However, this file was added
to the incorrect path, and does not follow the pattern of the rest of
the repository (to put all application files under src/). This commit
amends the location of this file, as well as the modman installer
reference, to be consistent with the rest of the files.