ISSUE-22 refactor (Manifest Model): Use the _construct method instead of __construct
Traditionally in OOP the goal is that each entity of an object should
have a predictable behaviour based on it's constructed parameters. This
achieved by performing logic on the creation of the object, and not
allowing the modification of this set state once the object has been
created (that is, not allowing "re-initialisation")

PHP provides the magic method __construct() to do this, which is simply
invoked on object construction. However, Magento uses this method to
perform it's own bootstrap. In the class that's extended (Varien_Object)
another method, _construct(), is provided to allow additional
bootstrapping while not overriding the parent's bootstrap behaviour.

This commit swaps the use of __construct() to _construct(), to preserve
parent behaviour while still bootstrapping correctly.