ISSUE-31 refactor (Configuration): Change default display to "standalone"
The current default value is "browser". When submitted, the author felt
this to be a reasonable default, as it is the last fallback of the
application, and thus likely to be widely supported. However, this logic
has now been revised.

When users install the extension, it's likely they are expecting it to
contribute some way to the site being an "installable application".
Chrome's audit tool (lighthouse) indicates that for it to be
installable, it must be one of

  - fullscreen
  - minimal-ui
  - standalone

After consultation, it was determined that standalone is a better
default.