AD-HOC feat (Configuration): Add a field that allows user image
Currently, the only way to modify the image used is to modify the system
configuration to reference a new image specified in the skin directory.
However, it stands to reason that, given an admin can modify the store
logo image, they should also be able to modify the manifest logo image.

This commit adds configuration that uses the default Magento image
handler to upload a logo image to the media directory, such that it can
be later discovered and used for a logo.

This design was taken verbatem from the following:

https://magento.stackexchange.com/a/87380/12719

Design Notes:

no usage:

  Another stream of work is implementing the logic to both use and
  resize this image